Parse response body as JSON only if it's non-empty
An empty string is not valid JSON, and `JSON.parse` fails on such input.
TransferWise (TW) had been recently returning empty bodies with their
Strong Customer Response (SCA) challenge responses for a few days,
presumably because headers alone are sufficient for passing SCA.

This caused the response parsing code in this library to fail, since it
was expecting a valid JSON response in the body. TW fixed this on their
end, but this commit should help prevent similar errors in the future.